01-CON | Refactoring some elements with using RangeV3
