+++ .travis.yml
@@ -1,0 +1,35 @@
+os: Linux
+dist: xenial
+sudo: false #disable sudo in scripts
+language: python
+python: "3.7"
+
+compiler: clang
+
+addons:
+  apt:
+    sources:
+    - ubuntu-toolchain-r-test
+    - sourceline: 'deb https://apt.llvm.org/xenial/ llvm-toolchain-xenial-9 main'
+      key_url: 'https://apt.llvm.org/llvm-snapshot.gpg.key'
+    packages:
+    - clang-9
+    - cmake
+    - cmake-data 
+env: TOOLSET=clang COMPILER=clang++-9 CXXSTD=c++17 TEST_SUITE=misc
+
+before_script:
+    - pip install conan
+    - conan user
+    - export CXX="clang++-9"
+    - export CC="clang-9"
+
+script:
+  - mkdir conanBuild/
+  - cd conanBuild/
+  - conan install ..
+  - cd ..
+  - cmake .
+  - cmake --build .
+  - cmake --build . --target test
+  - cmake --build . --target package
