+++ CMakeLists.txt
@@ -1,10 +1,13 @@
 cmake_minimum_required(VERSION 3.12)
 
+
 set(MAJOR_VERSION 0)
 set(MINOR_VERSION 1)
+
 set(ENV{TRAVIS_BUILD_NUMBER} 1) #plug
 
 project(ip_filter_csv VERSION ${MAJOR_VERSION}.${MINOR_VERSION}.$ENV{TRAVIS_BUILD_NUMBER})
+set(CMAKE_CXX_STANDARD 17)
 
 configure_file(version.vers ${PROJECT_SOURCE_DIR}/version.h)
