+++ libSource/filters.tpp
@@ -5,8 +5,8 @@
 
 namespace libSource
 {
-   
     template<typename ... Args>
+    [[deprecated]]
     vectVectInt filter(vectVectInt & ipBytes, const Args ... args)
     {
         vectVectInt ipSorted;
@@ -44,4 +44,39 @@ namespace libSource
     
     }
 
+    template<typename Type>
+    void show(std::vector<Type> const& showing)
+    {
+       int val = 0; 
+       
+       ranges::for_each(showing, [&val](auto element)
+       {
+           char symbol = ' ';
+
+           if(val < 3) symbol = '.'; 
+           else { symbol = ' '; }
+           
+           val++;
+           std::cout << element << symbol;
+           
+
+       });
+    }
+
+    template<typename Type>
+    void show(Type const& showing)
+    {
+        std::cout << showing << '\n';
+    }
+
+    template<typename Type>
+    void showAllContainer(std::vector<std::vector<Type>> container)
+    {
+       ranges::for_each(container, 
+       [](auto element)
+       {
+           show(element);
+           std::cout << '\n';
+       });
+    }
 };
