+++ libUnitTest/libUnitTest.h
@@ -1,12 +1,6 @@
 #pragma once
-#include "libUnitTest.h"
+
 #include "../version.h"
 
 extern int globArgc; 
 extern char **globArgv;
-
-namespace UnitTests
-{
-    bool checkBuildVersion();
-    int checkRegexp();
-};
